Add files via upload
Changes:
*added loading percentages in LichessProject.py
*NEW FILE: Main.py
removed all the testing lines at the bottom of LichessProject.py and transferred to Main.py so all the testing code is separate from the \\
code that's doing the work. All testing should be done from Main now
*NEW FILE: AlgorithmAnalysis.py - uses another new file Opening.py for better structure
contains algorithms(currently one) to analyze the chess data
*added a print function to Game.py for debugging purposes, currently not called anywhere

I think thats all the changes
Tried to fix the issue that came up when I fixed the time bottleneck but couldn't. Still might be possible but reverted to the original, functional version for now